chainlit-app/requirements.txt
chainlit>=1.0.0
fastapi>=0.100.0
websockets>=14.1
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0